fix: prevent resolver from widening output type
If the resolver for a field returns a union where one of the constituents
satisfies the expected type, then TypeScript widens the `Out` type to allow the
union, but the resolver's type should be limited to what the `type` option
allows. Adding this `NoInfer` utility prevents TypeScript from widening the
type in this case.